Create a new type for CM parameters.
This is a first commit in a series that aims at making the code more
modular and "less intelligent". In particular, cm_class is to be split
into cm_param, holding the CM parameters, and a smaller cm_class holding
the results of class polynomial computations. In this commit, the fields
of cm_param are simply copied into cm_class during cm_class_init.

* lib/param.c: New file.
* lib/Makefile.am: Register new file.
* lib/class.c (doubleeta_compute_parameter): Move content to...
* lib/param.c (doubleeta_compute_parameter): This function with an
  adapted prototype.
* lib/class.c (cm_class_compute_parameter): Move content to...
* lib/param.c (cm_param_init): This new function.
* lib/cm.h (__cm_param_struct, cm_param_t, cm_param_ptr, cm_param_srcptr):
  New types.
  (cm_param_init): Declare function prototype.
  (cm_class_init): Adapt function declaration.
* lib/class.c (cm_class_init): Adapt to new prototype, taking a cm_param
  instead of an invariant and a discriminant.
* lib/jmodp.c (cm_class_get_j_mod_P), src/classpol.c (main): Call
  cm_param_init before cm_class_init.