Treat discriminants in batches.
There is a very moderate effect on the time for trial division, which is
not a dominating step. So good parameter selection is difficult, since
the final running time varies a lot when different discriminants are
chosen in the middle. Also choosing the optimal discriminant in each batch
leads to slightly more primality tests.

* lib/ecpp.c (compute_cardinalities): Correct a bug when no_d > 1.
  (find_ecpp_discriminant): Implement treating discriminants in batches.