Compute the number of new signed primes differently.
Instead of adding a fixed number of signed primes depending on the size of
the number to be certified, compute the exact number of signed primes to
add so that the expected number of suitable curve cardinalities is
sufficiently large: at least 5 in the first round, or 1 later.
This should in particular benefit numbers for which it is difficult to
find a suitable prime: The previous strategy added signed primes in such
small quantities that the expected number of curves could be virtually 0,
leading to many loops containing expensive trial divisions.

* lib/ecpp.c (expected_no_curves): New function.
  (compute_sorted_discriminants): Drop arguments.
  (find_ecpp_discriminant): Drop arguments. Compute a suitable number of
  signed primes to add.
  (cm_ecpp1): Adapt function call.