Allow the subproduct tree for gcds to become a forest.
In this way, if the accumulated size of the m_j becomes larger than the n
for which n mod m_j is computed, we do not lose time in the construction
of the subproduct tree of the m_j. This is a first step towards splitting
n over several processes in the MPI version.

* lib/ecpp.c (cm_mpz_tree_gcd): Extend the code to work with subproduct
  forests.