Use higher trial division bounds in ECPP with MPI.
This requires the primorial to be split over more MPI processes, and thus
reduces the splitting of the numbers to be trial divided. Trial division
becomes slower, but for very large numbers, this is not the dominant step.
The first step on 640 cores with nextprime(10^40000) takes 1970s (of which
539s for trial division and 1026s for primality tests), instead of 1582s
(34s, 1150s) with the previous version. However, the step also gains
58 bits instead of 42 bits. So per gained bit, this is slightly faster,
and in any case it should result in a shorter certificate.

* lib/cm-impl.h (cm_mpi_get_tree_gcd): Add smoothness bound B to function
  prototype.
* lib/mpi.c (no_prim_opt): Remove static variable.
  (compute_no_prim): New function.
  (cm_mpi_get_tree_gcd): Change prototype and call new function.
  (mpi_worker): Call new function.
* lib/ecpp.c (trial_div): Change prototype for the MPI version to add
  smoothness bound B.
  (contains_ecpp_discriminant, trial_div, find_ecpp_discriminant): Adapt
  function calls.
  (ecpp1): Compute B differently for the MPI version.