Do not require a minimal gain for the parallel version.
This is expected to speed up proofs for difficult primes; instead of
increasing the number of qstar beyond reason, we sidestep the issue by
hoping to find an easier next prime of similar size. The certificate
size is expected to increase, but since the parallel version treats many
curves at once and chooses the one with the best gain, the difference
should be small in most cases with sufficient parallelism.

* lib/ecpp.c (ecpp1): Set delta to 2 in the parallel version.